Moved test logic to test.go
